fix properties_controller

+++ db/schema.rb
@@ -31,8 +31,8 @@ ActiveRecord::Schema[7.0].define(version: 2023_10_13_031749) do
 
   create_table "property_addresses", force: :cascade do |t|
     t.string "name"
-    t.bigint "latitude", null: false
-    t.bigint "longitude", null: false
+    t.float "latitude", null: false
+    t.float "longitude", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
